create sorted lookup tables & faster lookup table creation Issue #3
The varlab and date lookup tables that are based on is.csv now keep their order
as they appear in is.csv which mirrors basicly the order of Secutrials Datensatztabelle and the order
of variables as they appear in the data tables. Now, they come ordered when created.

The value lookup table which is based on cl.csv is sorted after creation, because cl.csv is ordered
alphabeticly and not by the secutrial data structure.

The functions that create the lookup list are now running much faster and therefore the whole reading process
has been speeded up.
	modified:   R/read-tsExport.R